+++ extension/src/shared/plateau/layers/TilesetFeatureIndex.ts
