+++ extension/src/shared/plateau/layers/index.ts
@@ -1,3 +1,3 @@
 export * from "./3dtiles";
-export * from "./TileFeatureIndex";
+export * from "./TilesetFeatureIndex";
 export * from "./TilesetProperties";
